books crud fixes complete.

+++ backend/src/helpers/helper.ts
@@ -1,4 +1,4 @@
-import {promises as fsPromises} from "fs";
+import {promises as fsPromises, existsSync} from "fs";
 
 export const generateRandomString = (length) => {
     let result = '';
@@ -21,7 +21,7 @@ export const getFileExtension = (filename) => {
 }
 
 export const uploadFile = async (dir_path, file_path, file) => {
-    await fsPromises.mkdir(dir_path, { recursive: true });
+    await fsPromises.mkdir(dir_path, {recursive: true});
     await fsPromises.writeFile(file_path, file.buffer);
 }
 
@@ -29,10 +29,11 @@ export const deleteFileFromUploads = async (app_url, delete_path) => {
     if (delete_path.includes('uploads') && delete_path.includes(app_url)) {
         delete_path = delete_path.replace(app_url, '.');
         try {
-            fsPromises.access(delete_path)
-                .then(() => {
-                    return fsPromises.unlink(delete_path);
-                });
+            if (existsSync(delete_path))
+                fsPromises.access(delete_path)
+                    .then(() => {
+                        return fsPromises.unlink(delete_path);
+                    });
         } catch (error) {
             console.log('error');
         }
